fix: Correct project_id foreign key column type to unsignedBigInteger in repositories migration

+++ database/migrations/2024_08_04_094159_create_repositories_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('repositories', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('project_id');
+            $table->unsignedBigInteger('project_id');
             $table->foreign('project_id')->references('id')->on('projects')->onDelete('cascade');
             $table->string('name');
             $table->string('repo_path');
